moving files into vqe_playgroud module and introducing vqe-playground entry point
