+++ vqe_playground/command_line.py
@@ -1,0 +1,5 @@
+from .vqe_main import VQEPlayground
+
+
+def main():
+    VQEPlayground().main()
