+++ vqe_playground/data/bell_histogram.png
