+++ vqe_playground/utils/colors.py
