+++ vqe_playground/utils/data/bell_circuit.png
