+++ vqe_playground/utils/data/bell_histogram.png
