+++ vqe_playground/utils/data/bell_qsphere.png
