+++ vqe_playground/utils/data/gate_images/ctrl_gate.png
