+++ vqe_playground/utils/data/gate_images/ctrl_gate_bottom_wire.png
