+++ vqe_playground/utils/data/gate_images/ctrl_gate_top_wire.png
