+++ vqe_playground/utils/data/gate_images/h_gate.png
