+++ vqe_playground/utils/data/gate_images/iden_gate.png
