+++ vqe_playground/utils/data/gate_images/not_gate.png
