+++ vqe_playground/utils/data/gate_images/not_gate_above_ctrl.png
