+++ vqe_playground/utils/data/gate_images/rx_gate.png
