+++ vqe_playground/utils/data/gate_images/ry_gate.png
