+++ vqe_playground/utils/data/gate_images/rz_gate.png
