+++ vqe_playground/utils/data/gate_images/s_gate.png
