+++ vqe_playground/utils/data/gate_images/sdg_gate.png
