+++ vqe_playground/utils/data/gate_images/swap_gate.png
