+++ vqe_playground/utils/data/gate_images/t_gate.png
