+++ vqe_playground/utils/data/gate_images/tdg_gate.png
