+++ vqe_playground/utils/data/gate_images/trace_gate.png
