+++ vqe_playground/utils/data/gate_images/x_gate.png
