+++ vqe_playground/utils/data/gate_images/y_gate.png
