+++ vqe_playground/utils/data/gate_images/z_gate.png
