+++ vqe_playground/utils/data/images/circuit-grid-cursor-60px.png
