+++ vqe_playground/utils/data/images/circuit-grid-cursor.png
