+++ vqe_playground/utils/data/mel_circ.png
