+++ vqe_playground/viz/expectation_grid.py
