--- vqe_playground_pkg/command_line.py
@@ -1,6 +1,0 @@
-# from main_module import vqe_main
-from main_module.vqe_main import VQEPlayground
-
-
-def main():
-    VQEPlayground().main()
